Fiks etter pr
