+++ src/main/web_src/src/components/bestillingsveileder/stegVelger/steg/steg3/Steg3.js
@@ -19,7 +19,7 @@ export const Steg3 = ({ formikBag }) => {
 				bestillingsdata={formikBag.values}
 				heading="Hvilke testmiljø vil du opprette personene i?"
 			/>
-			<MalForm formikBag={formikBag} />
+			{!formikBag.values.hasOwnProperty('organisasjon') && <MalForm formikBag={formikBag} />}
 		</div>
 	)
 }
